requirements.txt
bleach
duckduckgo-search
googlesearch-python
gradio[oauth]
markdownify
mwparserfromhell
requests
smolagents
tinycss2
wikipedia-api